requirements.txt
pydantic>=2.7.2
pyspark>=3.5.1
delta-spark
testcontainers==4.5.1
pytest-mock
delta-spark
findspark
psutil